Messing around with JSON.
It's now the case that the standbys calendar only loads after an AJAX call
returns from receiveing JSON from the `updates_view' config, which uses the JSON
renderer to echo back JSON. `updates_view' will eventually echo back all days
standbys are scheduled for in the current month. The only concern is that no
calendar gets displayed until this query happens, which may be too slow for
everyone's taste.

The alternative is to embed the standby days in HTML and have Javascript go to
work, parsing these days, which may not be a bad alternative.