Pushed up statistics collected by git.
In lieu of Dr. Ladd's announcement that there will be no final and more of our
grade will be based on this final project, I have used a free tool called
git-stats that takes information collected by git and presents it in a readable
way.

To view these statistics open git-stats/index.html in a web browser.