Did some work with standbys.
Clicking on a day on the calendar which and standby is scheduled will bring up a
list of standbys that are happening that day. When one of these standbys pages
are visited, the user will be given the opportunity to sign up. If the user
signs up, then they have the ability to request coverage.