Cut down on duplication with Events and Standby calendars.
Now instead of events.pt and standbys.pt having their own identical code for the
calendar, they both use the code in /static/calendar.js. The minor differences
in behavior expected in the Events calendar vs. the Standbys calendar are
accounted for by events.pt and standbys.pt declaring global `episode' variables
so the calendar code knows which calendar it's suppose to be displaying.

I also downloaded the jQuery javascript and CSS files that were being fetched
from the jQuery website every time the website was being loaded and stuck them
in the static directory. The javascript and CSS files are now only included in
events.pt and standbys.pt instead of template.pt, also, cutting down on excess
inclusion of these files.