Got the duty crew calendar working correctly.
Now when you click on a date on the duty crew calendar, you are redirected to a
separate page that lists everyone on that night. If you are working and on call
that night, there is a button that allows you to request coverage. Likewise, if
you are on call and have already requested coverage, a button is displayed that
allows you to retract your request.