Calendar now only queries for the month it is displaying.
Going from month to month is a little bit slower now because every month change
requires a query to the Pyramid server for all of the events/standbys/duty crews
for that month, but it is much more space efficient.

Events and Standbys that show up as lists have their times printed in a much
nicer format now, too.